Merging purgatory into osm_app branch

+++ iOS/G3MiOSSDK/Commons/Geometry/Box.cpp
@@ -102,8 +102,8 @@ Vector2F Box::projectedExtent(const G3MRenderContext* rc) const {
   float lowerY = pixel0._y;
   float upperY = pixel0._y;
 
-  const int cornersSize = corners.size();
-  for (int i = 1; i < cornersSize; i++) {
+  const size_t cornersSize = corners.size();
+  for (size_t i = 1; i < cornersSize; i++) {
     const Vector2F pixel = currentCamera->point2Pixel(corners[i]);
 
     const float x = pixel._x;
@@ -137,8 +137,8 @@ double Box::projectedArea(const G3MRenderContext* rc) const {
   float lowerY = pixel0._y;
   float upperY = pixel0._y;
 
-  const int cornersSize = corners.size();
-  for (int i = 1; i < cornersSize; i++) {
+  const size_t cornersSize = corners.size();
+  for (size_t i = 1; i < cornersSize; i++) {
     const Vector2F pixel = currentCamera->point2Pixel(corners[i]);
 
     const float x = pixel._x;
